feat(tunnelify-provider): Client and Provider can now use tokens to reuse tunnels names
Token are exchanged through Socket.IO query. If the client can't find a
token saved in ~/.tunnelify with the corresponding mount path or the
provider can't regognize the token, a new token will be provided by
provider or the client simply will not pass the token.
In case a token
is passed from the client, the requested name will be ignored.